refactor: split mailbox delivery watcher helpers
